% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_color_filter.R
\name{any_color_filter}
\alias{any_color_filter}
\title{any color filter}
\usage{
any_color_filter(
  vector_for_filter,
  any_color = c(155, 0, 155),
  blend_factor = 0.5
)
}
\description{
Uses a vector_for_filter generated from your image by the apply_any_color_filter function.
Generates a raw color vector that will be processed into an image by the apply_any_color_filter function.
this function is built into another function. No need to use by itself.
}
