% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_any_color_filter.R
\name{apply_any_color_filter}
\alias{apply_any_color_filter}
\title{Applies the any color filter}
\usage{
apply_any_color_filter(
  image_path,
  any_color = c(255, 0, 255),
  blend_factor = 0.5
)
}
\arguments{
\item{image_path}{# an object to which you have assigned the path or name (if in WD) to the image you want to filter.}

\item{any_color}{= c(255, 0, 255) # Choose three RBG values (0-255) to customize the color}
}
\value{
a filtered image converted back to a bitmap
}
\description{
Loads the image from your directory and converts it into a bitmap and the vector_for_filter
required for the filter functions. Returns a final_image with your applied filter which can
be printed or saved. Uses both the filter function and the convert to bitmap function.
}
\examples{
# example code

library(magick)
image_path <- "Path/to/your/image_1.jpg"
filtered_image <- apply_any_color_filter(image_path)
# Will generate an "filtered_image" object which can be printed and saved.
print(filtered_image) # View the final image
image_write(filtered_image, path = "my_image.jpg", format = "jpeg") # Save the final image as a JPEG.

}
