% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_grayscale_filter.R
\name{apply_grayscale_filter}
\alias{apply_grayscale_filter}
\title{Applies the grayscale filter to a grayscale image}
\usage{
apply_grayscale_filter(image_path)
}
\arguments{
\item{image_path}{# an object to which you have assigned the path or name (if in WD) to the image you want to filter.}
}
\value{
a filtered image converted back to a bitmap
}
\description{
Loads the image from your directory and converts it into a bitmap and the vector_for_filter
required for the filter functions. Returns a final_image with your applied filter which can
be printed or saved. Uses both the filter function and the convert to bitmap function.
}
\examples{
image_path <- "Path/to/your/image_1.jpg"
filtered_image <- apply_grayscale_filter(image_path)
# Will generate an "filtered_image" object which can be printed and saved.
print(filtered_image) # View the final image
image_write(filtered_image, path = "my_image.jpg", format = "jpeg") # Save the final image as a JPEG.
}
