% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGB_color_filter.R
\name{color_filter}
\alias{color_filter}
\title{Red, green and blue filter}
\usage{
color_filter(vector_for_filter, color = "red")
}
\description{
Uses a vector_for_filter generated from your image by the apply_RGB_color_filter function.
Generates a raw RGB vector that will be processed into an image by the apply_RGB_color_filter function.
this function is built into another function. No need to use by itself.
}
