% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grayscale_filter.R
\name{grayscale_filter}
\alias{grayscale_filter}
\title{Grayscale filter}
\usage{
grayscale_filter(vector_for_filter)
}
\description{
Uses a vector_for_filter generated from your image by the apply_grayscale_filter function.
Generates a raw grayscale vector that will be processed into an image by the apply_grayscale_filter_function.
this function is built into another function. No need to use by itself.
}
